requirements.txt
asgiref==3.2.3
Babel==2.8.0
Django==3.0.3
django-phonenumber-field==4.0.0
phonenumbers==8.11.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
